refactor(approval-gate): move public-API tests to tests/*.rs
Final test split. The 86 remaining inline tests in lib.rs that
exercise only the crate's `pub` surface now live in tests/*.rs,
organized by the area they cover. The state-machine proptest gets
its own file. lib.rs drops from 2369 to 52 lines (just module
declarations + re-exports).

New tests/ layout:
- common/mod.rs    — shared fakes: FakeExecutor, InMemoryStateBus,
                     FailingStateBus, sample_call(), empty_policy_rules()
- lifecycle.rs     — 19 tests: build_pending_record, transition_record,
                     maybe_flip_timed_out, collect_timed_out_for_sweep,
                     is_terminal_status, pending_key
- wire.rs          — 10 tests: extract_call, block_reply_for,
                     requires_approval
- intercept.rs     — 9 tests: handle_intercept (replays, fail-closed,
                     session_id stamping, force_pending)
- resolve.rs       — 19 tests: handle_resolve, cascade-on-`always`,
                     handle_lookup_record
- delivery.rs      — 25 tests: list_pending / list_undelivered /
                     ack_delivered / consume_undelivered /
                     flush_delivered / sweep_session
- misc.rs          — 4 tests: FN_* constants, unverified_marker_targets,
                     FakeExecutor smoke test
- state_machine.rs — the proptest with the four lifecycle invariants

Each tests/*.rs uses `mod common;` for shared fakes — only one source
of truth for in-memory bus/executor stand-ins.

141 approval-gate tests pass across 13 test binaries (previously 6).
turn-orchestrator + harness + shell verified green.
