+++ approval-gate/src/lib.rs
@@ -49,2321 +49,3 @@ use intercept::{
 use state::{merge_from_approval_marker_if_needed, rule_for};
 #[cfg(test)]
 use sweeper::timeout_resolved_event;
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::{json, Value};
-    use std::sync::{Arc, RwLock};
-
-    /// Empty policy ruleset for tests that exercise [`handle_resolve`]
-    /// without cascading. Each call freshly constructs the lock so unit
-    /// tests stay independent — there's no shared mutable state.
-    fn empty_policy_rules() -> std::sync::RwLock<crate::rules::Ruleset> {
-        std::sync::RwLock::new(crate::rules::Ruleset::new())
-    }
-
-    #[test]
-    fn maybe_flip_timed_out_returns_some_when_pending_and_expired() {
-        let rec = build_pending_record("tc-1", "shell::fs::write", &json!({}), 1_000, 60_000);
-        let flipped = maybe_flip_timed_out(&rec, 70_000).expect("should flip");
-        assert_eq!(flipped["status"], "timed_out");
-        // Timeout carries no Denial — the status alone explains the outcome.
-        assert!(flipped.get("denial").is_none());
-        assert!(flipped.get("decision_reason").is_none());
-    }
-
-    #[test]
-    fn maybe_flip_timed_out_returns_none_when_pending_and_not_expired() {
-        let rec = build_pending_record("tc-1", "shell::fs::write", &json!({}), 1_000, 60_000);
-        assert!(maybe_flip_timed_out(&rec, 60_000).is_none());
-        assert!(maybe_flip_timed_out(&rec, 1_500).is_none());
-    }
-
-    #[test]
-    fn maybe_flip_timed_out_returns_none_when_not_pending() {
-        let rec = json!({
-            "function_call_id": "tc-1",
-            "status": "executed",
-            "expires_at": 1_000_u64,
-        });
-        assert!(maybe_flip_timed_out(&rec, 999_999_999).is_none());
-    }
-
-    #[test]
-    fn transition_record_stamps_resolved_at_for_terminal_status() {
-        let base = build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000);
-        let rec = transition_record_with_now(
-            &base,
-            "executed",
-            Some(json!({"ok": true})),
-            None,
-            None,
-            12_345,
-        );
-        assert_eq!(rec["resolved_at"].as_u64(), Some(12_345));
-    }
-
-    #[test]
-    fn transition_record_preserves_existing_resolved_at_on_relift() {
-        let base = build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000);
-        let first = transition_record_with_now(
-            &base,
-            "executed",
-            Some(json!({"ok": true})),
-            None,
-            None,
-            12_345,
-        );
-        let second = transition_record_with_now(
-            &first,
-            "executed",
-            Some(json!({"ok": true})),
-            None,
-            None,
-            99_999,
-        );
-        assert_eq!(second["resolved_at"].as_u64(), Some(12_345));
-    }
-
-    #[test]
-    fn transition_record_does_not_stamp_resolved_at_for_intermediate_status() {
-        let base = build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000);
-        let rec =
-            transition_record_with_now(&base, "approved", None, None, None, 12_345);
-        assert!(rec.get("resolved_at").is_none());
-    }
-
-    #[tokio::test]
-    async fn handle_list_undelivered_caps_at_default_limit_and_reports_omitted() {
-        let bus = InMemoryStateBus::new();
-        for i in 0..75 {
-            let cid = format!("c{i}");
-            let mut rec = transition_record_with_now(
-                &build_pending_record(&cid, "shell::fs::write", &json!({}), 1_000, 60_000),
-                "executed",
-                Some(json!({"ok": true})),
-                None,
-                None,
-                1_000 + i as u64,
-            );
-            rec.as_object_mut()
-                .unwrap()
-                .insert("session_id".into(), Value::String("s1".into()));
-            bus.set(STATE_SCOPE, &pending_key("s1", &cid), rec)
-                .await
-                .unwrap();
-        }
-        let resp =
-            handle_list_undelivered(&bus, STATE_SCOPE, json!({"session_id": "s1"}), 100_000).await;
-        assert_eq!(resp["entries"].as_array().unwrap().len(), 50);
-        assert_eq!(resp["omitted"].as_u64(), Some(25));
-    }
-
-    #[tokio::test]
-    async fn handle_list_undelivered_honors_explicit_limit() {
-        let bus = InMemoryStateBus::new();
-        for i in 0..10 {
-            let cid = format!("c{i}");
-            let mut rec = transition_record_with_now(
-                &build_pending_record(&cid, "shell::fs::write", &json!({}), 1_000, 60_000),
-                "executed",
-                Some(json!({"ok": true})),
-                None,
-                None,
-                1_000 + i as u64,
-            );
-            rec.as_object_mut()
-                .unwrap()
-                .insert("session_id".into(), Value::String("s1".into()));
-            bus.set(STATE_SCOPE, &pending_key("s1", &cid), rec)
-                .await
-                .unwrap();
-        }
-        let resp = handle_list_undelivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s1", "limit": 3}),
-            100_000,
-        )
-        .await;
-        assert_eq!(resp["entries"].as_array().unwrap().len(), 3);
-        assert_eq!(resp["omitted"].as_u64(), Some(7));
-    }
-
-    #[tokio::test]
-    async fn handle_list_undelivered_returns_oldest_first_by_resolved_at() {
-        let bus = InMemoryStateBus::new();
-        for (i, ts) in [(0_u32, 5_000_u64), (1, 1_000), (2, 3_000)] {
-            let cid = format!("c{i}");
-            let mut rec = transition_record_with_now(
-                &build_pending_record(&cid, "shell::fs::write", &json!({}), 1_000, 60_000),
-                "executed",
-                Some(json!({"ok": true})),
-                None,
-                None,
-                ts,
-            );
-            rec.as_object_mut()
-                .unwrap()
-                .insert("session_id".into(), Value::String("s1".into()));
-            bus.set(STATE_SCOPE, &pending_key("s1", &cid), rec)
-                .await
-                .unwrap();
-        }
-        let resp = handle_list_undelivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s1", "limit": 10}),
-            100_000,
-        )
-        .await;
-        let entries = resp["entries"].as_array().unwrap();
-        let ids: Vec<&str> = entries
-            .iter()
-            .map(|e| e["function_call_id"].as_str().unwrap())
-            .collect();
-        assert_eq!(ids, vec!["c1", "c2", "c0"]);
-    }
-
-    #[tokio::test]
-    async fn handle_list_undelivered_omitted_is_zero_when_under_limit() {
-        let bus = InMemoryStateBus::new();
-        let mut rec = transition_record_with_now(
-            &build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000),
-            "executed",
-            Some(json!({"ok": true})),
-            None,
-            None,
-            1_500,
-        );
-        rec.as_object_mut()
-            .unwrap()
-            .insert("session_id".into(), Value::String("s1".into()));
-        bus.set(STATE_SCOPE, &pending_key("s1", "c1"), rec)
-            .await
-            .unwrap();
-        let resp =
-            handle_list_undelivered(&bus, STATE_SCOPE, json!({"session_id": "s1"}), 100_000).await;
-        assert_eq!(resp["entries"].as_array().unwrap().len(), 1);
-        assert_eq!(resp["omitted"].as_u64(), Some(0));
-    }
-
-    #[tokio::test]
-    async fn handle_consume_undelivered_stamps_returned_entries() {
-        let bus = InMemoryStateBus::new();
-        for i in 0..3 {
-            let cid = format!("c{i}");
-            let mut rec = transition_record_with_now(
-                &build_pending_record(&cid, "shell::fs::write", &json!({}), 1_000, 60_000),
-                "executed",
-                Some(json!({"ok": true})),
-                None,
-                None,
-                1_000 + i as u64,
-            );
-            rec.as_object_mut()
-                .unwrap()
-                .insert("session_id".into(), Value::String("s1".into()));
-            bus.set(STATE_SCOPE, &pending_key("s1", &cid), rec)
-                .await
-                .unwrap();
-        }
-        let resp = handle_consume_undelivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s1", "turn_id": "turn-7", "limit": 10}),
-            100_000,
-        )
-        .await;
-        assert_eq!(resp["ok"], json!(true));
-        assert_eq!(resp["entries"].as_array().unwrap().len(), 3);
-        assert_eq!(resp["omitted"].as_u64(), Some(0));
-        let next =
-            handle_list_undelivered(&bus, STATE_SCOPE, json!({"session_id": "s1"}), 100_000).await;
-        assert_eq!(next["entries"].as_array().unwrap().len(), 0);
-    }
-
-    #[tokio::test]
-    async fn handle_consume_undelivered_respects_limit_and_leaves_remainder() {
-        let bus = InMemoryStateBus::new();
-        for i in 0..5 {
-            let cid = format!("c{i}");
-            let mut rec = transition_record_with_now(
-                &build_pending_record(&cid, "shell::fs::write", &json!({}), 1_000, 60_000),
-                "executed",
-                Some(json!({"ok": true})),
-                None,
-                None,
-                1_000 + i as u64,
-            );
-            rec.as_object_mut()
-                .unwrap()
-                .insert("session_id".into(), Value::String("s1".into()));
-            bus.set(STATE_SCOPE, &pending_key("s1", &cid), rec)
-                .await
-                .unwrap();
-        }
-        let resp = handle_consume_undelivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s1", "turn_id": "turn-7", "limit": 2}),
-            100_000,
-        )
-        .await;
-        assert_eq!(resp["entries"].as_array().unwrap().len(), 2);
-        assert_eq!(resp["omitted"].as_u64(), Some(3));
-        let next =
-            handle_list_undelivered(&bus, STATE_SCOPE, json!({"session_id": "s1"}), 100_000).await;
-        assert_eq!(next["entries"].as_array().unwrap().len(), 3);
-    }
-
-    #[tokio::test]
-    async fn handle_consume_undelivered_missing_turn_id_returns_error() {
-        let bus = InMemoryStateBus::new();
-        let resp = handle_consume_undelivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s1"}),
-            100_000,
-        )
-        .await;
-        assert_eq!(resp["ok"], json!(false));
-        assert_eq!(resp["error"], json!("missing_turn_id"));
-    }
-
-    #[tokio::test]
-    async fn handle_flush_delivered_stamps_all_unacked_terminals() {
-        let bus = InMemoryStateBus::new();
-        for i in 0..5 {
-            let cid = format!("c{i}");
-            let mut rec = transition_record_with_now(
-                &build_pending_record(&cid, "shell::fs::write", &json!({}), 1_000, 60_000),
-                "executed",
-                Some(json!({"ok": true})),
-                None,
-                None,
-                1_000 + i as u64,
-            );
-            rec.as_object_mut()
-                .unwrap()
-                .insert("session_id".into(), Value::String("s1".into()));
-            bus.set(STATE_SCOPE, &pending_key("s1", &cid), rec)
-                .await
-                .unwrap();
-        }
-        let resp = handle_flush_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s1", "turn_id": "manual-flush"}),
-        )
-        .await;
-        assert_eq!(resp["ok"], json!(true));
-        assert_eq!(resp["stamped"].as_u64(), Some(5));
-        let next =
-            handle_list_undelivered(&bus, STATE_SCOPE, json!({"session_id": "s1"}), 100_000).await;
-        assert_eq!(next["entries"].as_array().unwrap().len(), 0);
-    }
-
-    #[tokio::test]
-    async fn handle_flush_delivered_skips_pending_records() {
-        let bus = InMemoryStateBus::new();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "c1"),
-            build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000),
-        )
-        .await
-        .unwrap();
-        let resp = handle_flush_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s1", "turn_id": "manual-flush"}),
-        )
-        .await;
-        assert_eq!(resp["stamped"].as_u64(), Some(0));
-        let still = bus
-            .get(STATE_SCOPE, &pending_key("s1", "c1"))
-            .await
-            .unwrap();
-        assert_eq!(still["status"].as_str(), Some("pending"));
-        assert!(still.get("delivered_in_turn_id").is_none());
-    }
-
-    #[tokio::test]
-    async fn handle_flush_delivered_idempotent_on_already_stamped() {
-        let bus = InMemoryStateBus::new();
-        let mut rec = transition_record_with_now(
-            &build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000),
-            "executed",
-            Some(json!({"ok": true})),
-            None,
-            None,
-            1_500,
-        );
-        {
-            let obj = rec.as_object_mut().unwrap();
-            obj.insert(
-                "delivered_in_turn_id".into(),
-                Value::String("turn-prev".into()),
-            );
-            obj.insert("session_id".into(), Value::String("s1".into()));
-        }
-        bus.set(STATE_SCOPE, &pending_key("s1", "c1"), rec)
-            .await
-            .unwrap();
-        let resp = handle_flush_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s1", "turn_id": "manual-flush"}),
-        )
-        .await;
-        assert_eq!(resp["stamped"].as_u64(), Some(0));
-        let still = bus
-            .get(STATE_SCOPE, &pending_key("s1", "c1"))
-            .await
-            .unwrap();
-        assert_eq!(still["delivered_in_turn_id"].as_str(), Some("turn-prev"));
-    }
-
-    #[tokio::test]
-    async fn handle_list_undelivered_returns_terminal_records_with_no_delivered_stamp() {
-        let bus = InMemoryStateBus::new();
-        let mut r1 = transition_record(
-            &build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000),
-            "executed",
-            Some(json!({"ok": true})),
-            None,
-            None,
-        );
-        r1.as_object_mut()
-            .unwrap()
-            .insert("session_id".into(), Value::String("s1".into()));
-        bus.set(STATE_SCOPE, &pending_key("s1", "c1"), r1)
-            .await
-            .unwrap();
-        let mut r2 = transition_record(
-            &build_pending_record("c2", "shell::fs::write", &json!({}), 1_000, 60_000),
-            "denied",
-            None,
-            None,
-            Some(Denial::UserCorrected {
-                feedback: "nope".into(),
-            }),
-        );
-        r2.as_object_mut()
-            .unwrap()
-            .insert("session_id".into(), Value::String("s1".into()));
-        bus.set(STATE_SCOPE, &pending_key("s1", "c2"), r2)
-            .await
-            .unwrap();
-
-        let resp =
-            handle_list_undelivered(&bus, STATE_SCOPE, json!({"session_id": "s1"}), 100_000).await;
-        let entries = resp["entries"].as_array().unwrap();
-        assert_eq!(entries.len(), 2);
-        assert_eq!(resp["omitted"].as_u64(), Some(0));
-    }
-
-    #[tokio::test]
-    async fn handle_list_undelivered_excludes_pending_records() {
-        let bus = InMemoryStateBus::new();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "c1"),
-            build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000),
-        )
-        .await
-        .unwrap();
-
-        let resp =
-            handle_list_undelivered(&bus, STATE_SCOPE, json!({"session_id": "s1"}), 1_500).await;
-        assert_eq!(resp["entries"].as_array().unwrap().len(), 0);
-    }
-
-    #[tokio::test]
-    async fn handle_list_undelivered_empty_session_returns_empty() {
-        let bus = InMemoryStateBus::new();
-        let resp =
-            handle_list_undelivered(&bus, STATE_SCOPE, json!({"session_id": "s1"}), 1_500).await;
-        assert_eq!(resp["entries"], json!([]));
-    }
-
-    #[tokio::test]
-    async fn handle_list_undelivered_excludes_records_stamped_with_delivered_turn_id() {
-        let bus = InMemoryStateBus::new();
-        let mut rec = transition_record(
-            &build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000),
-            "executed",
-            Some(json!({"ok": true})),
-            None,
-            None,
-        );
-        {
-            let obj = rec.as_object_mut().unwrap();
-            obj.insert(
-                "delivered_in_turn_id".into(),
-                Value::String("turn-prev".into()),
-            );
-            obj.insert("session_id".into(), Value::String("s1".into()));
-        }
-        bus.set(STATE_SCOPE, &pending_key("s1", "c1"), rec)
-            .await
-            .unwrap();
-
-        let mut r2 = transition_record(
-            &build_pending_record("c2", "shell::fs::write", &json!({}), 1_000, 60_000),
-            "executed",
-            Some(json!({"ok": true})),
-            None,
-            None,
-        );
-        r2.as_object_mut()
-            .unwrap()
-            .insert("session_id".into(), Value::String("s1".into()));
-        bus.set(STATE_SCOPE, &pending_key("s1", "c2"), r2)
-            .await
-            .unwrap();
-
-        let resp =
-            handle_list_undelivered(&bus, STATE_SCOPE, json!({"session_id": "s1"}), 100_000).await;
-        let entries = resp["entries"].as_array().unwrap();
-        assert_eq!(entries.len(), 1);
-        assert_eq!(entries[0]["function_call_id"], "c2");
-    }
-
-    #[tokio::test]
-    async fn handle_list_undelivered_returns_empty_when_session_id_missing() {
-        let bus = InMemoryStateBus::new();
-        let resp = handle_list_undelivered(&bus, STATE_SCOPE, json!({}), 1_500).await;
-        assert_eq!(resp["entries"], json!([]));
-    }
-
-    #[tokio::test]
-    async fn handle_ack_delivered_stamps_records_with_turn_id() {
-        let bus = InMemoryStateBus::new();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "c1"),
-            transition_record(
-                &build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000),
-                "executed",
-                Some(json!({"ok": true})),
-                None,
-                None,
-            ),
-        )
-        .await
-        .unwrap();
-
-        let resp = handle_ack_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({
-                "session_id": "s1",
-                "call_ids": ["c1"],
-                "turn_id": "turn-1",
-            }),
-        )
-        .await;
-        assert_eq!(resp["ok"], json!(true));
-        assert_eq!(resp["stamped"], json!(1));
-
-        let rec = bus
-            .get(STATE_SCOPE, &pending_key("s1", "c1"))
-            .await
-            .unwrap();
-        assert_eq!(rec["delivered_in_turn_id"], "turn-1");
-    }
-
-    #[tokio::test]
-    async fn handle_ack_delivered_is_idempotent_keeps_first_turn_id() {
-        let bus = InMemoryStateBus::new();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "c1"),
-            transition_record(
-                &build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000),
-                "executed",
-                Some(json!({"ok": true})),
-                None,
-                None,
-            ),
-        )
-        .await
-        .unwrap();
-
-        let _ = handle_ack_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({
-                "session_id": "s1", "call_ids": ["c1"], "turn_id": "turn-first",
-            }),
-        )
-        .await;
-        let resp = handle_ack_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({
-                "session_id": "s1", "call_ids": ["c1"], "turn_id": "turn-second",
-            }),
-        )
-        .await;
-        assert_eq!(resp["stamped"], json!(0), "second ack must not re-stamp");
-
-        let rec = bus
-            .get(STATE_SCOPE, &pending_key("s1", "c1"))
-            .await
-            .unwrap();
-        assert_eq!(rec["delivered_in_turn_id"], "turn-first");
-    }
-
-    #[tokio::test]
-    async fn handle_ack_delivered_skips_unknown_call_ids_silently() {
-        let bus = InMemoryStateBus::new();
-        let resp = handle_ack_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({
-                "session_id": "s1", "call_ids": ["ghost"], "turn_id": "turn-1",
-            }),
-        )
-        .await;
-        assert_eq!(resp["ok"], json!(true));
-        assert_eq!(resp["stamped"], json!(0));
-    }
-
-    #[tokio::test]
-    async fn handle_resolve_on_expired_pending_flips_to_timed_out_and_ignores_decision() {
-        let bus = InMemoryStateBus::new();
-        let exec = FakeExecutor::default();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "tc-1"),
-            build_pending_record("tc-1", "shell::fs::write", &json!({}), 1_000, 60_000),
-        )
-        .await
-        .unwrap();
-
-        let resp = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &empty_policy_rules(),
-            json!({"session_id":"s1","function_call_id":"tc-1","decision":"allow"}),
-            70_000,
-        )
-        .await;
-        assert_eq!(resp["ok"], json!(false));
-        assert_eq!(resp["error"], "timed_out");
-
-        assert!(exec.calls.lock().unwrap().is_empty());
-
-        let rec = bus
-            .get(STATE_SCOPE, &pending_key("s1", "tc-1"))
-            .await
-            .unwrap();
-        assert_eq!(rec["status"], "timed_out");
-    }
-
-    #[test]
-    fn fn_constants_match_spec_strings() {
-        assert_eq!(FN_RESOLVE, "approval::resolve");
-        assert_eq!(FN_LIST_PENDING, "approval::list_pending");
-        assert_eq!(FN_LIST_UNDELIVERED, "approval::list_undelivered");
-        assert_eq!(FN_ACK_DELIVERED, "approval::ack_delivered");
-        assert_eq!(FN_LOOKUP_RECORD, "approval::lookup_record");
-    }
-
-    #[test]
-    fn is_terminal_status_returns_true_for_terminal_states() {
-        assert!(is_terminal_status("executed"));
-        assert!(is_terminal_status("failed"));
-        assert!(is_terminal_status("denied"));
-        assert!(is_terminal_status("timed_out"));
-    }
-
-    #[test]
-    fn is_terminal_status_returns_false_for_in_progress_states() {
-        assert!(!is_terminal_status("pending"));
-        assert!(!is_terminal_status("approved"));
-        assert!(!is_terminal_status("anything_else"));
-        assert!(!is_terminal_status(""));
-    }
-
-    #[test]
-    fn pending_key_includes_session_and_tool_call_id() {
-        assert_eq!(pending_key("s1", "tc-1"), "s1/tc-1");
-    }
-
-    #[test]
-    fn extract_call_reads_session_id_and_function_call_from_envelope() {
-        let envelope = json!({
-            "event_id": "evt-1",
-            "reply_stream": "rs-1",
-            "payload": {
-                "function_call": { "id": "tc-1", "function_id": "write", "arguments": {"path": "/tmp/x"} },
-                "approval_required": ["write"],
-                "session_id": "s1",
-            }
-        });
-        let call = extract_call(&envelope).expect("decoded");
-        assert_eq!(call.session_id, "s1");
-        assert_eq!(call.function_call_id, "tc-1");
-        assert_eq!(call.function_id, "write");
-        assert_eq!(call.event_id, "evt-1");
-        assert_eq!(call.reply_stream, "rs-1");
-        assert!(call.approval_required.iter().any(|s| s == "write"));
-    }
-
-    #[test]
-    fn extract_call_accepts_legacy_tool_call_envelope_with_name() {
-        let envelope = json!({
-            "event_id": "evt-1",
-            "reply_stream": "rs-1",
-            "payload": {
-                "tool_call": { "id": "tc-1", "name": "write", "arguments": {} },
-                "approval_required": ["write"],
-                "session_id": "s1",
-            }
-        });
-        let call = extract_call(&envelope).expect("decoded");
-        assert_eq!(call.function_call_id, "tc-1");
-        assert_eq!(call.function_id, "write");
-    }
-
-    #[test]
-    fn requires_approval_only_for_listed_functions() {
-        let call = IncomingCall {
-            session_id: "s1".into(),
-            function_call_id: "tc-1".into(),
-            function_id: "ls".into(),
-            args: json!({}),
-            approval_required: vec!["write".into()],
-            event_id: "e".into(),
-            reply_stream: "r".into(),
-        };
-        assert!(!call.requires_approval());
-
-        let call2 = IncomingCall {
-            function_id: "write".into(),
-            ..call
-        };
-        assert!(call2.requires_approval());
-    }
-
-    #[test]
-    fn build_pending_record_sets_status_and_expiry() {
-        let now = 1_000_000;
-        let rec = build_pending_record("tc-1", "write", &json!({"x": 1}), now, 60_000);
-        assert_eq!(rec["status"], "pending");
-        assert_eq!(rec["function_call_id"], "tc-1");
-        assert_eq!(rec["expires_at"], 1_060_000);
-    }
-
-    #[test]
-    fn block_reply_for_decision_allow_does_not_block() {
-        let reply = block_reply_for(&Decision::Allow);
-        assert_eq!(reply["block"], false);
-    }
-
-    #[test]
-    fn block_reply_for_deny_emits_structured_denial() {
-        let reply = block_reply_for(&Decision::Deny(Denial::UserRejected));
-        assert_eq!(reply["block"], true);
-        assert_eq!(reply["denial"]["kind"], "user_rejected");
-        assert!(reply.as_object().unwrap().get("reason").is_none());
-    }
-
-    #[test]
-    fn block_reply_for_policy_deny_carries_classifier_detail() {
-        let reply = block_reply_for(&Decision::Deny(Denial::Policy {
-            classifier_reason: "command matches denylist".into(),
-            classifier_fn: "shell::classify_argv".into(),
-        }));
-        assert_eq!(reply["block"], true);
-        assert_eq!(reply["denial"]["kind"], "policy");
-        assert_eq!(
-            reply["denial"]["detail"]["classifier_reason"],
-            "command matches denylist"
-        );
-        assert_eq!(
-            reply["denial"]["detail"]["classifier_fn"],
-            "shell::classify_argv"
-        );
-    }
-
-    #[test]
-    fn block_reply_for_user_corrected_carries_feedback() {
-        let reply = block_reply_for(&Decision::Deny(Denial::UserCorrected {
-            feedback: "use git diff instead".into(),
-        }));
-        assert_eq!(reply["denial"]["kind"], "user_corrected");
-        assert_eq!(
-            reply["denial"]["detail"]["feedback"],
-            "use git diff instead"
-        );
-    }
-
-    #[test]
-    fn extract_call_returns_none_when_function_call_absent() {
-        let envelope = json!({
-            "event_id": "evt-1",
-            "reply_stream": "rs-1",
-            "payload": { "session_id": "s1", "approval_required": ["write"] }
-        });
-        assert!(extract_call(&envelope).is_none());
-    }
-
-    #[test]
-    fn extract_call_returns_none_when_session_id_absent() {
-        let envelope = json!({
-            "event_id": "evt-1",
-            "reply_stream": "rs-1",
-            "payload": {
-                "tool_call": { "id": "tc-1", "name": "write", "arguments": {} }
-            }
-        });
-        assert!(extract_call(&envelope).is_none());
-    }
-
-    #[test]
-    fn block_reply_for_allow_omits_denial_and_reason() {
-        let reply = block_reply_for(&Decision::Allow);
-        assert_eq!(reply["block"], false);
-        assert!(
-            reply.get("reason").is_none(),
-            "Allow must not include reason: {reply}"
-        );
-        assert!(
-            reply.get("denial").is_none(),
-            "Allow must not include denial: {reply}"
-        );
-    }
-
-    use std::sync::Mutex;
-
-    fn sample_call() -> IncomingCall {
-        IncomingCall {
-            session_id: "s1".into(),
-            function_call_id: "tc-1".into(),
-            function_id: "shell::fs::write".into(),
-            args: json!({"path": "/tmp/a"}),
-            approval_required: vec!["shell::fs::write".into()],
-            event_id: "evt-1".into(),
-            reply_stream: "rs-1".into(),
-        }
-    }
-
-    #[tokio::test]
-    async fn handle_intercept_returns_pending_envelope_when_call_is_gated() {
-        let bus = InMemoryStateBus::new();
-        let call = sample_call();
-        let reply = handle_intercept(&bus, STATE_SCOPE, &call, 1_000, 60_000, false).await;
-        assert_eq!(reply["block"], json!(true));
-        assert_eq!(reply["status"], json!("pending"));
-        assert_eq!(reply["call_id"], json!("tc-1"));
-        assert_eq!(reply["function_id"], json!("shell::fs::write"));
-        // Pending status is self-describing — no `reason` or `denial` field
-        // is emitted while the call is in-flight.
-        assert!(reply.get("reason").is_none());
-        assert!(reply.get("denial").is_none());
-    }
-
-    #[tokio::test]
-    async fn handle_intercept_writes_pending_record_to_state() {
-        let bus = InMemoryStateBus::new();
-        let call = sample_call();
-        let _ = handle_intercept(&bus, STATE_SCOPE, &call, 1_000, 60_000, false).await;
-        let key = pending_key(&call.session_id, &call.function_call_id);
-        let rec = bus
-            .get(STATE_SCOPE, &key)
-            .await
-            .expect("pending record written");
-        assert_eq!(rec["status"], "pending");
-        assert_eq!(rec["function_call_id"], "tc-1");
-        assert_eq!(rec["expires_at"], 61_000);
-    }
-
-    #[tokio::test]
-    async fn handle_intercept_passes_through_when_call_is_not_gated() {
-        let bus = InMemoryStateBus::new();
-        let mut call = sample_call();
-        call.approval_required = vec!["other".into()];
-        let reply = handle_intercept(&bus, STATE_SCOPE, &call, 1_000, 60_000, false).await;
-        assert_eq!(reply["block"], json!(false));
-        let key = pending_key(&call.session_id, &call.function_call_id);
-        assert!(
-            bus.get(STATE_SCOPE, &key).await.is_none(),
-            "no record written"
-        );
-    }
-
-    #[tokio::test]
-    async fn handle_intercept_force_pending_writes_when_not_on_required_list() {
-        let bus = InMemoryStateBus::new();
-        let mut call = sample_call();
-        call.approval_required = vec!["other".into()];
-        let reply = handle_intercept(&bus, STATE_SCOPE, &call, 1_000, 60_000, true).await;
-        assert_eq!(reply["block"], json!(true));
-        assert_eq!(reply["status"], json!("pending"));
-        let key = pending_key(&call.session_id, &call.function_call_id);
-        assert!(bus.get(STATE_SCOPE, &key).await.is_some());
-    }
-
-    #[tokio::test]
-    async fn handle_lookup_record_returns_null_when_missing() {
-        let bus = InMemoryStateBus::new();
-        let v = handle_lookup_record(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s1", "function_call_id": "c1"}),
-        )
-        .await;
-        assert!(v.is_null());
-    }
-
-    #[tokio::test]
-    async fn handle_lookup_record_returns_record_when_present() {
-        let bus = InMemoryStateBus::new();
-        let call = sample_call();
-        let _ = handle_intercept(&bus, STATE_SCOPE, &call, 1_000, 60_000, false).await;
-        let v = handle_lookup_record(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s1", "function_call_id": "tc-1"}),
-        )
-        .await;
-        assert_eq!(v["status"], json!("pending"));
-        assert_eq!(v["function_id"], json!("shell::fs::write"));
-    }
-
-    #[derive(Default)]
-    struct FakeExecutor {
-        calls: Mutex<Vec<(String, Value, String, String)>>,
-        response: Mutex<Option<Result<Value, String>>>,
-    }
-
-    #[async_trait::async_trait]
-    impl FunctionExecutor for FakeExecutor {
-        async fn invoke(
-            &self,
-            function_id: &str,
-            args: Value,
-            function_call_id: &str,
-            session_id: &str,
-        ) -> Result<Value, String> {
-            self.calls.lock().unwrap().push((
-                function_id.to_string(),
-                args,
-                function_call_id.to_string(),
-                session_id.to_string(),
-            ));
-            self.response
-                .lock()
-                .unwrap()
-                .clone()
-                .unwrap_or_else(|| Ok(json!({"ok": true})))
-        }
-    }
-
-    #[tokio::test]
-    async fn handle_resolve_allow_invokes_function_and_records_executed() {
-        let bus = InMemoryStateBus::new();
-        let exec = FakeExecutor::default();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "tc-1"),
-            build_pending_record(
-                "tc-1",
-                "shell::fs::write",
-                &json!({"path":"/a"}),
-                1_000,
-                60_000,
-            ),
-        )
-        .await
-        .unwrap();
-
-        let resp = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &empty_policy_rules(),
-            json!({
-                "session_id": "s1",
-                "function_call_id": "tc-1",
-                "decision": "allow",
-            }),
-            1_500,
-        )
-        .await;
-        assert_eq!(resp["ok"], json!(true));
-
-        let calls = exec.calls.lock().unwrap().clone();
-        assert_eq!(calls.len(), 1);
-        assert_eq!(calls[0].0, "shell::fs::write");
-        assert_eq!(calls[0].1, json!({"path":"/a"}));
-        assert_eq!(calls[0].2, "tc-1");
-        assert_eq!(calls[0].3, "s1");
-
-        let rec = bus
-            .get(STATE_SCOPE, &pending_key("s1", "tc-1"))
-            .await
-            .unwrap();
-        assert_eq!(rec["status"], "executed");
-        assert_eq!(rec["result"], json!({"ok": true}));
-    }
-
-    #[tokio::test]
-    async fn allow_without_always_does_not_cascade() {
-        // Two pending shell::exec calls in the same session. Resolving
-        // the first with allow (always=false) must NOT touch the second.
-        let bus = InMemoryStateBus::new();
-        let exec = FakeExecutor::default();
-        for cid in ["tc-1", "tc-2"] {
-            let mut rec = build_pending_record(cid, "shell::exec", &json!({}), 1_000, 60_000);
-            rec.as_object_mut()
-                .unwrap()
-                .insert("session_id".into(), json!("s1"));
-            bus.set(STATE_SCOPE, &pending_key("s1", cid), rec)
-                .await
-                .unwrap();
-        }
-        let rules = empty_policy_rules();
-        let resp = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &rules,
-            json!({
-                "session_id": "s1",
-                "function_call_id": "tc-1",
-                "decision": "allow",
-            }),
-            1_500,
-        )
-        .await;
-        assert_eq!(resp["ok"], true);
-        assert!(
-            resp.get("cascaded").is_none(),
-            "cascaded field must be omitted when always was not set: {resp}"
-        );
-        let other = bus
-            .get(STATE_SCOPE, &pending_key("s1", "tc-2"))
-            .await
-            .unwrap();
-        assert_eq!(other["status"], "pending");
-        assert_eq!(rules.read().unwrap().len(), 0, "rule must not be pushed");
-    }
-
-    #[tokio::test]
-    async fn allow_with_always_pushes_rule_and_cascades_same_session_pending() {
-        // Three pending calls in session s1: two shell::exec, one
-        // shell::fs::write. Resolving the first shell::exec with
-        // always=true must:
-        //   1. Push an Allow rule for shell::exec
-        //   2. Auto-resolve the other shell::exec pending in this session
-        //   3. Leave the shell::fs::write pending untouched
-        let bus = InMemoryStateBus::new();
-        let exec = FakeExecutor::default();
-        for (cid, fn_id) in [
-            ("tc-1", "shell::exec"),
-            ("tc-2", "shell::exec"),
-            ("tc-3", "shell::fs::write"),
-        ] {
-            let mut rec = build_pending_record(cid, fn_id, &json!({}), 1_000, 60_000);
-            rec.as_object_mut()
-                .unwrap()
-                .insert("session_id".into(), json!("s1"));
-            bus.set(STATE_SCOPE, &pending_key("s1", cid), rec)
-                .await
-                .unwrap();
-        }
-        let rules = empty_policy_rules();
-
-        let resp = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &rules,
-            json!({
-                "session_id": "s1",
-                "function_call_id": "tc-1",
-                "decision": "allow",
-                "always": true,
-            }),
-            1_500,
-        )
-        .await;
-        assert_eq!(resp["ok"], true);
-        assert_eq!(
-            resp["cascaded"], json!(1),
-            "tc-2 should cascade; tc-1 originator excluded; tc-3 not matched"
-        );
-
-        // The Allow rule for shell::exec is now in the shared ruleset.
-        let pushed = rules.read().unwrap();
-        assert_eq!(pushed.len(), 1);
-        assert_eq!(pushed[0].permission, "shell::exec");
-        assert_eq!(pushed[0].action, rules::Action::Allow);
-        drop(pushed);
-
-        // Originator and cascaded record both transitioned to executed.
-        let r1 = bus
-            .get(STATE_SCOPE, &pending_key("s1", "tc-1"))
-            .await
-            .unwrap();
-        let r2 = bus
-            .get(STATE_SCOPE, &pending_key("s1", "tc-2"))
-            .await
-            .unwrap();
-        let r3 = bus
-            .get(STATE_SCOPE, &pending_key("s1", "tc-3"))
-            .await
-            .unwrap();
-        assert_eq!(r1["status"], "executed");
-        assert_eq!(r2["status"], "executed");
-        assert_eq!(
-            r3["status"], "pending",
-            "non-matching function_id must stay pending: {r3}"
-        );
-
-        // Executor was invoked twice: originator + cascaded.
-        assert_eq!(exec.calls.lock().unwrap().len(), 2);
-    }
-
-    #[tokio::test]
-    async fn cascade_does_not_cross_session_boundary() {
-        // tc-1 in session s1, tc-2 in session s2 — both shell::exec.
-        // Resolving s1/tc-1 with always must not touch s2/tc-2.
-        let bus = InMemoryStateBus::new();
-        let exec = FakeExecutor::default();
-        for (session, cid) in [("s1", "tc-1"), ("s2", "tc-2")] {
-            let mut rec = build_pending_record(cid, "shell::exec", &json!({}), 1_000, 60_000);
-            rec.as_object_mut()
-                .unwrap()
-                .insert("session_id".into(), json!(session));
-            bus.set(STATE_SCOPE, &pending_key(session, cid), rec)
-                .await
-                .unwrap();
-        }
-        let rules = empty_policy_rules();
-
-        let resp = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &rules,
-            json!({
-                "session_id": "s1",
-                "function_call_id": "tc-1",
-                "decision": "allow",
-                "always": true,
-            }),
-            1_500,
-        )
-        .await;
-        assert_eq!(resp["ok"], true);
-        assert!(
-            resp.get("cascaded").is_none() || resp["cascaded"] == json!(0),
-            "no record in s1 to cascade onto; tc-2 in s2 must NOT be touched: {resp}"
-        );
-
-        let other_session = bus
-            .get(STATE_SCOPE, &pending_key("s2", "tc-2"))
-            .await
-            .unwrap();
-        assert_eq!(other_session["status"], "pending");
-        assert_eq!(
-            exec.calls.lock().unwrap().len(),
-            1,
-            "only the originator should have been invoked"
-        );
-    }
-
-    #[tokio::test]
-    async fn cascade_skips_originator_record() {
-        // Single pending record. always=true must not double-resolve it.
-        let bus = InMemoryStateBus::new();
-        let exec = FakeExecutor::default();
-        let mut rec = build_pending_record("tc-1", "shell::exec", &json!({}), 1_000, 60_000);
-        rec.as_object_mut()
-            .unwrap()
-            .insert("session_id".into(), json!("s1"));
-        bus.set(STATE_SCOPE, &pending_key("s1", "tc-1"), rec)
-            .await
-            .unwrap();
-        let rules = empty_policy_rules();
-
-        let resp = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &rules,
-            json!({
-                "session_id": "s1",
-                "function_call_id": "tc-1",
-                "decision": "allow",
-                "always": true,
-            }),
-            1_500,
-        )
-        .await;
-        assert_eq!(resp["ok"], true);
-        // Originator counts under the existing allow path, not the cascade.
-        assert!(resp.get("cascaded").is_none() || resp["cascaded"] == json!(0));
-        assert_eq!(exec.calls.lock().unwrap().len(), 1);
-    }
-
-    #[tokio::test]
-    async fn cascade_skips_already_resolved_records_in_session() {
-        // Two records in s1: tc-1 pending, tc-2 already terminal. The
-        // cascade must skip tc-2.
-        let bus = InMemoryStateBus::new();
-        let exec = FakeExecutor::default();
-        let mut r1 = build_pending_record("tc-1", "shell::exec", &json!({}), 1_000, 60_000);
-        r1.as_object_mut()
-            .unwrap()
-            .insert("session_id".into(), json!("s1"));
-        bus.set(STATE_SCOPE, &pending_key("s1", "tc-1"), r1)
-            .await
-            .unwrap();
-        let mut r2 = build_pending_record("tc-2", "shell::exec", &json!({}), 1_000, 60_000);
-        r2.as_object_mut()
-            .unwrap()
-            .insert("session_id".into(), json!("s1"));
-        let r2_done = transition_record(&r2, "executed", Some(json!({"ok": true})), None, None);
-        bus.set(STATE_SCOPE, &pending_key("s1", "tc-2"), r2_done)
-            .await
-            .unwrap();
-
-        let rules = empty_policy_rules();
-        let resp = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &rules,
-            json!({
-                "session_id": "s1",
-                "function_call_id": "tc-1",
-                "decision": "allow",
-                "always": true,
-            }),
-            1_500,
-        )
-        .await;
-        assert_eq!(resp["ok"], true);
-        // tc-2 is terminal — not pending — so cascade skips it.
-        assert!(resp.get("cascaded").is_none() || resp["cascaded"] == json!(0));
-    }
-
-    #[tokio::test]
-    async fn handle_resolve_deny_does_not_invoke_function() {
-        let bus = InMemoryStateBus::new();
-        let exec = FakeExecutor::default();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "tc-1"),
-            build_pending_record("tc-1", "shell::fs::write", &json!({}), 1_000, 60_000),
-        )
-        .await
-        .unwrap();
-
-        let resp = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &empty_policy_rules(),
-            json!({
-                "session_id": "s1",
-                "function_call_id": "tc-1",
-                "decision": "deny",
-                "denial": {
-                    "kind": "user_corrected",
-                    "detail": { "feedback": "not authorized" }
-                },
-            }),
-            1_500,
-        )
-        .await;
-        assert_eq!(resp["ok"], json!(true));
-
-        assert!(exec.calls.lock().unwrap().is_empty());
-
-        let rec = bus
-            .get(STATE_SCOPE, &pending_key("s1", "tc-1"))
-            .await
-            .unwrap();
-        assert_eq!(rec["status"], "denied");
-        assert_eq!(rec["denial"]["kind"], "user_corrected");
-        assert_eq!(rec["denial"]["detail"]["feedback"], "not authorized");
-    }
-
-    #[tokio::test]
-    async fn handle_resolve_allow_records_failed_when_function_errors() {
-        let bus = InMemoryStateBus::new();
-        let exec = FakeExecutor::default();
-        *exec.response.lock().unwrap() = Some(Err("EACCES".into()));
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "tc-1"),
-            build_pending_record("tc-1", "shell::fs::write", &json!({}), 1_000, 60_000),
-        )
-        .await
-        .unwrap();
-
-        let resp = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &empty_policy_rules(),
-            json!({"session_id":"s1","function_call_id":"tc-1","decision":"allow"}),
-            1_500,
-        )
-        .await;
-        assert_eq!(resp["ok"], json!(true));
-
-        let rec = bus
-            .get(STATE_SCOPE, &pending_key("s1", "tc-1"))
-            .await
-            .unwrap();
-        assert_eq!(rec["status"], "failed");
-        assert_eq!(rec["error"], "EACCES");
-    }
-
-    #[tokio::test]
-    async fn fake_executor_records_calls() {
-        let exec = FakeExecutor::default();
-        let out = exec
-            .invoke("shell::fs::write", json!({"x": 1}), "cid", "sid")
-            .await
-            .unwrap();
-        assert_eq!(out, json!({"ok": true}));
-        let calls = exec.calls.lock().unwrap().clone();
-        assert_eq!(calls.len(), 1);
-        assert_eq!(calls[0].0, "shell::fs::write");
-        assert_eq!(calls[0].2, "cid");
-        assert_eq!(calls[0].3, "sid");
-    }
-
-    struct InMemoryStateBus {
-        store: Mutex<std::collections::HashMap<String, Value>>,
-    }
-
-    impl InMemoryStateBus {
-        fn new() -> Self {
-            Self {
-                store: Mutex::new(std::collections::HashMap::new()),
-            }
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl StateBus for InMemoryStateBus {
-        async fn set(&self, scope: &str, key: &str, value: Value) -> Result<(), iii_sdk::IIIError> {
-            self.store
-                .lock()
-                .unwrap()
-                .insert(format!("{scope}/{key}"), value);
-            Ok(())
-        }
-        async fn get(&self, scope: &str, key: &str) -> Option<Value> {
-            self.store
-                .lock()
-                .unwrap()
-                .get(&format!("{scope}/{key}"))
-                .cloned()
-        }
-        async fn list_prefix(&self, scope: &str, prefix: &str) -> Vec<Value> {
-            let map = self.store.lock().unwrap();
-            map.iter()
-                .filter(|(k, _)| k.starts_with(&format!("{scope}/{prefix}")))
-                .map(|(_, v)| v.clone())
-                .collect()
-        }
-    }
-
-    #[tokio::test]
-    async fn resolve_flips_status_when_pending() {
-        let bus = InMemoryStateBus::new();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "tc-1"),
-            build_pending_record("tc-1", "write", &json!({}), 0, 60_000),
-        )
-        .await
-        .unwrap();
-
-        let exec = FakeExecutor::default();
-        let out = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &empty_policy_rules(),
-            json!({
-                "function_call_id": "tc-1",
-                "session_id": "s1",
-                "decision": "allow",
-            }),
-            1_500,
-        )
-        .await;
-
-        assert_eq!(out["ok"], true);
-        let stored = bus
-            .get(STATE_SCOPE, &pending_key("s1", "tc-1"))
-            .await
-            .unwrap();
-        assert_eq!(stored["status"], "executed");
-    }
-
-    #[tokio::test]
-    async fn resolve_accepts_legacy_tool_call_id_field() {
-        let bus = InMemoryStateBus::new();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "tc-1"),
-            build_pending_record("tc-1", "write", &json!({}), 0, 60_000),
-        )
-        .await
-        .unwrap();
-
-        let exec = FakeExecutor::default();
-        let out = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &empty_policy_rules(),
-            json!({
-                "tool_call_id": "tc-1",
-                "session_id": "s1",
-                "decision": "allow",
-            }),
-            1_500,
-        )
-        .await;
-
-        assert_eq!(out["ok"], true);
-    }
-
-    #[tokio::test]
-    async fn resolve_rejects_already_resolved_entry() {
-        let bus = InMemoryStateBus::new();
-        let mut rec = build_pending_record("tc-1", "write", &json!({}), 0, 60_000);
-        rec["status"] = json!("allow");
-        bus.set(STATE_SCOPE, &pending_key("s1", "tc-1"), rec)
-            .await
-            .unwrap();
-
-        let exec = FakeExecutor::default();
-        let out = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &empty_policy_rules(),
-            json!({"function_call_id": "tc-1", "session_id": "s1", "decision": "deny"}),
-            1_500,
-        )
-        .await;
-        assert_eq!(out["ok"], false);
-        assert_eq!(out["error"], "already_resolved");
-    }
-
-    #[tokio::test]
-    async fn list_pending_returns_only_pending_for_session() {
-        let bus = InMemoryStateBus::new();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "tc-1"),
-            build_pending_record("tc-1", "write", &json!({}), 0, 60_000),
-        )
-        .await
-        .unwrap();
-        let mut resolved = build_pending_record("tc-2", "write", &json!({}), 0, 60_000);
-        resolved["status"] = json!("allow");
-        bus.set(STATE_SCOPE, &pending_key("s1", "tc-2"), resolved)
-            .await
-            .unwrap();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("other", "tc-3"),
-            build_pending_record("tc-3", "write", &json!({}), 0, 60_000),
-        )
-        .await
-        .unwrap();
-
-        let out = handle_list_pending(&bus, STATE_SCOPE, json!({ "session_id": "s1" })).await;
-        let items = out["pending"].as_array().unwrap();
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0]["function_call_id"], "tc-1");
-    }
-
-    #[tokio::test]
-    async fn resolve_deny_without_denial_defaults_to_user_rejected() {
-        let bus = InMemoryStateBus::new();
-        let _ = bus
-            .set(
-                STATE_SCOPE,
-                &pending_key("s1", "tc-1"),
-                build_pending_record("tc-1", "write", &json!({}), 0, 60_000),
-            )
-            .await;
-
-        let exec = FakeExecutor::default();
-        let out = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &empty_policy_rules(),
-            json!({
-                "session_id": "s1",
-                "function_call_id": "tc-1",
-                "decision": "deny",
-            }),
-            1_500,
-        )
-        .await;
-        assert_eq!(out["ok"], true);
-
-        let stored = bus
-            .get(STATE_SCOPE, &pending_key("s1", "tc-1"))
-            .await
-            .unwrap();
-        assert_eq!(stored["status"], "denied");
-        assert_eq!(stored["denial"]["kind"], "user_rejected");
-    }
-
-    #[tokio::test]
-    async fn resolve_deny_rejects_malformed_denial() {
-        let bus = InMemoryStateBus::new();
-        let _ = bus
-            .set(
-                STATE_SCOPE,
-                &pending_key("s1", "tc-1"),
-                build_pending_record("tc-1", "write", &json!({}), 0, 60_000),
-            )
-            .await;
-
-        let exec = FakeExecutor::default();
-        let out = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &empty_policy_rules(),
-            json!({
-                "session_id": "s1",
-                "function_call_id": "tc-1",
-                "decision": "deny",
-                "denial": { "kind": "not_a_real_kind" },
-            }),
-            1_500,
-        )
-        .await;
-        assert_eq!(out["ok"], false);
-        assert_eq!(out["error"], "bad_denial");
-    }
-
-    #[test]
-    fn transition_record_to_executed_attaches_result() {
-        let base = build_pending_record(
-            "tc-1",
-            "shell::fs::write",
-            &json!({"path":"/a"}),
-            1_000,
-            60_000,
-        );
-        let rec = transition_record(&base, "executed", Some(json!({"ok": true})), None, None);
-        assert_eq!(rec["status"], "executed");
-        assert_eq!(rec["result"], json!({"ok": true}));
-        assert!(rec.get("error").is_none() || rec["error"].is_null());
-        assert_eq!(rec["function_call_id"], "tc-1");
-        assert_eq!(rec["function_id"], "shell::fs::write");
-    }
-
-    #[test]
-    fn transition_record_to_failed_attaches_error() {
-        let base = build_pending_record("tc-1", "shell::fs::write", &json!({}), 1_000, 60_000);
-        let rec = transition_record(&base, "failed", None, Some("EACCES".into()), None);
-        assert_eq!(rec["status"], "failed");
-        assert_eq!(rec["error"], "EACCES");
-        assert!(rec.get("result").is_none() || rec["result"].is_null());
-    }
-
-    #[test]
-    fn transition_record_to_denied_attaches_structured_denial() {
-        let base = build_pending_record("tc-1", "shell::fs::write", &json!({}), 1_000, 60_000);
-        let rec = transition_record(
-            &base,
-            "denied",
-            None,
-            None,
-            Some(Denial::Policy {
-                classifier_reason: "not authorized".into(),
-                classifier_fn: "shell::classify_argv".into(),
-            }),
-        );
-        assert_eq!(rec["status"], "denied");
-        assert_eq!(rec["denial"]["kind"], "policy");
-        assert_eq!(rec["denial"]["detail"]["classifier_reason"], "not authorized");
-        assert!(
-            rec.get("decision_reason").is_none(),
-            "legacy decision_reason must not be written: {rec}"
-        );
-    }
-
-    #[test]
-    fn transition_record_to_timed_out_carries_no_denial() {
-        // Timeout status is self-describing — no Denial attached.
-        let base = build_pending_record("tc-1", "shell::fs::write", &json!({}), 1_000, 60_000);
-        let rec = transition_record(&base, "timed_out", None, None, None);
-        assert_eq!(rec["status"], "timed_out");
-        assert!(rec.get("denial").is_none());
-        assert!(rec.get("decision_reason").is_none());
-    }
-
-    #[test]
-    fn transition_record_preserves_delivered_in_turn_id_when_set() {
-        let mut base = build_pending_record("tc-1", "shell::fs::write", &json!({}), 1_000, 60_000);
-        base.as_object_mut().unwrap().insert(
-            "delivered_in_turn_id".into(),
-            Value::String("turn-X".into()),
-        );
-        let rec = transition_record(&base, "executed", Some(json!({"ok": true})), None, None);
-        assert_eq!(rec["delivered_in_turn_id"], "turn-X");
-    }
-
-    #[tokio::test]
-    async fn handle_sweep_session_flips_pending_records_to_timed_out() {
-        let bus = InMemoryStateBus::new();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "c1"),
-            build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000),
-        )
-        .await
-        .unwrap();
-
-        let resp = handle_sweep_session(&bus, STATE_SCOPE, json!({"session_id": "s1"})).await;
-        assert_eq!(resp["swept"], json!(1));
-
-        let rec = bus
-            .get(STATE_SCOPE, &pending_key("s1", "c1"))
-            .await
-            .unwrap();
-        assert_eq!(rec["status"], "timed_out");
-        // sweep_session no longer stamps a reason string — timed_out is
-        // self-describing per the Denial refactor.
-        assert!(rec.get("denial").is_none());
-        assert!(rec.get("decision_reason").is_none());
-    }
-
-    #[tokio::test]
-    async fn handle_sweep_session_ignores_legacy_reason_payload_field() {
-        // Old callers may still pass `reason` — approval-gate accepts the
-        // payload but does not persist it. Behavior is identical to a
-        // bare {session_id} payload.
-        let bus = InMemoryStateBus::new();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "c1"),
-            build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000),
-        )
-        .await
-        .unwrap();
-        let resp = handle_sweep_session(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s1", "reason": "run_stopped"}),
-        )
-        .await;
-        assert_eq!(resp["swept"], json!(1));
-        let rec = bus
-            .get(STATE_SCOPE, &pending_key("s1", "c1"))
-            .await
-            .unwrap();
-        assert_eq!(rec["status"], "timed_out");
-        assert!(rec.get("denial").is_none());
-    }
-
-    #[tokio::test]
-    async fn handle_sweep_session_skips_non_pending_records() {
-        let bus = InMemoryStateBus::new();
-        bus.set(
-            STATE_SCOPE,
-            &pending_key("s1", "c1"),
-            transition_record(
-                &build_pending_record("c1", "shell::fs::write", &json!({}), 1_000, 60_000),
-                "executed",
-                Some(json!({"ok": true})),
-                None,
-                None,
-            ),
-        )
-        .await
-        .unwrap();
-
-        let resp = handle_sweep_session(&bus, STATE_SCOPE, json!({"session_id": "s1"})).await;
-        assert_eq!(resp["swept"], json!(0));
-
-        let rec = bus
-            .get(STATE_SCOPE, &pending_key("s1", "c1"))
-            .await
-            .unwrap();
-        assert_eq!(rec["status"], "executed");
-    }
-
-    #[tokio::test]
-    async fn handle_sweep_session_returns_error_when_session_id_missing() {
-        let bus = InMemoryStateBus::new();
-        let resp = handle_sweep_session(&bus, STATE_SCOPE, json!({})).await;
-        assert_eq!(resp["ok"], json!(false));
-        assert_eq!(resp["error"], "missing_session_id");
-        assert_eq!(resp["swept"], json!(0));
-    }
-
-    // ── New reliability fixes ─────────────────────────────────────────────
-
-    /// A bus that always refuses writes, to exercise fail-closed semantics.
-    struct FailingStateBus;
-
-    #[async_trait::async_trait]
-    impl StateBus for FailingStateBus {
-        async fn set(
-            &self,
-            _scope: &str,
-            _key: &str,
-            _value: Value,
-        ) -> Result<(), iii_sdk::IIIError> {
-            Err(iii_sdk::IIIError::Runtime("kv unreachable".into()))
-        }
-        async fn get(&self, _scope: &str, _key: &str) -> Option<Value> {
-            None
-        }
-        async fn list_prefix(&self, _scope: &str, _prefix: &str) -> Vec<Value> {
-            Vec::new()
-        }
-    }
-
-    #[tokio::test]
-    async fn handle_intercept_fails_closed_on_state_write_error() {
-        let bus = FailingStateBus;
-        let call = sample_call();
-        let reply = handle_intercept(&bus, STATE_SCOPE, &call, 1_000, 60_000, false).await;
-        assert_eq!(
-            reply["block"],
-            json!(true),
-            "state write failure must NOT fail-open"
-        );
-        assert_eq!(reply["status"], json!("denied"));
-        assert_eq!(reply["denial"]["kind"], json!("state_error"));
-        assert_eq!(
-            reply["denial"]["detail"]["phase"],
-            json!("intercept_write_pending")
-        );
-        // The underlying error message is present but its exact text is
-        // bus-implementation-specific; just check it's non-empty.
-        assert!(
-            reply["denial"]["detail"]["error"]
-                .as_str()
-                .map(|s| !s.is_empty())
-                .unwrap_or(false),
-            "state_error detail must include error message: {reply}"
-        );
-        assert_eq!(reply["function_id"], json!("shell::fs::write"));
-    }
-
-    #[tokio::test]
-    async fn handle_intercept_stamps_session_id_into_pending_record() {
-        let bus = InMemoryStateBus::new();
-        let call = sample_call();
-        let _ = handle_intercept(&bus, STATE_SCOPE, &call, 1_000, 60_000, false).await;
-        let rec = bus
-            .get(
-                STATE_SCOPE,
-                &pending_key(&call.session_id, &call.function_call_id),
-            )
-            .await
-            .expect("pending record");
-        assert_eq!(rec["session_id"], json!(call.session_id));
-    }
-
-    #[test]
-    fn collect_timed_out_for_sweep_returns_expired_records_with_session_id() {
-        let mut rec = build_pending_record("tc-1", "shell::fs::write", &json!({}), 0, 60_000);
-        rec.as_object_mut()
-            .unwrap()
-            .insert("session_id".into(), json!("s-42"));
-        let pile = vec![
-            rec.clone(),
-            build_pending_record("tc-2", "shell::fs::write", &json!({}), 0, 999_999_999),
-        ];
-        let out = collect_timed_out_for_sweep(&pile, 70_000);
-        assert_eq!(out.len(), 1);
-        let (key, flipped, session_id, call_id) = &out[0];
-        assert_eq!(key, "s-42/tc-1");
-        assert_eq!(session_id, "s-42");
-        assert_eq!(call_id, "tc-1");
-        assert_eq!(flipped["status"], json!("timed_out"));
-        // Timeout carries no Denial — status is self-describing.
-        assert!(flipped.get("denial").is_none());
-        assert!(flipped.get("decision_reason").is_none());
-    }
-
-    #[test]
-    fn collect_timed_out_for_sweep_skips_records_without_session_id() {
-        // Legacy row (pre-session_id-stamping fix). The sweeper can't
-        // address the right session stream, so it must skip silently —
-        // lazy-flip on read will still pick it up.
-        let pile = vec![build_pending_record(
-            "tc-legacy",
-            "shell::fs::write",
-            &json!({}),
-            0,
-            60_000,
-        )];
-        let out = collect_timed_out_for_sweep(&pile, 70_000);
-        assert!(
-            out.is_empty(),
-            "legacy record without session_id must not be swept"
-        );
-    }
-
-    #[test]
-    fn unverified_marker_targets_lists_unasserted_rules() {
-        let rules = vec![
-            InterceptorRule {
-                function_id: "shell::exec".into(),
-                classifier: None,
-                classifier_timeout_ms: 2000,
-                inject_approval_marker: true,
-                marker_target_verified: false,
-            },
-            InterceptorRule {
-                function_id: "shell::exec_bg".into(),
-                classifier: None,
-                classifier_timeout_ms: 2000,
-                inject_approval_marker: true,
-                marker_target_verified: true,
-            },
-            InterceptorRule {
-                function_id: "no_marker::fn".into(),
-                classifier: None,
-                classifier_timeout_ms: 2000,
-                inject_approval_marker: false,
-                marker_target_verified: false,
-            },
-        ];
-        assert_eq!(unverified_marker_targets(&rules), vec!["shell::exec"]);
-    }
-
-    #[test]
-    fn unverified_marker_targets_empty_when_all_verified_or_marker_off() {
-        let rules = vec![
-            InterceptorRule {
-                function_id: "shell::exec".into(),
-                classifier: None,
-                classifier_timeout_ms: 2000,
-                inject_approval_marker: true,
-                marker_target_verified: true,
-            },
-            InterceptorRule {
-                function_id: "other".into(),
-                classifier: None,
-                classifier_timeout_ms: 2000,
-                inject_approval_marker: false,
-                marker_target_verified: false,
-            },
-        ];
-        assert!(unverified_marker_targets(&rules).is_empty());
-    }
-
-    // ── Boundary + edge-case tests prompted by cargo-mutants survivors ────
-    //
-    // Each test corresponds to a mutant the test suite previously didn't
-    // catch. Test name → mutated line in src/lib.rs.
-
-    #[tokio::test]
-    async fn handle_intercept_replay_of_terminal_record_returns_already_resolved() {
-        // mutant L331: replace `==` with `!=` in the replay defense — if
-        // flipped, terminal records would be overwritten with fresh pending.
-        let bus = InMemoryStateBus::new();
-        let call = sample_call();
-        let key = pending_key(&call.session_id, &call.function_call_id);
-        let terminal = transition_record(
-            &build_pending_record(
-                &call.function_call_id,
-                &call.function_id,
-                &call.args,
-                0,
-                60_000,
-            ),
-            "executed",
-            Some(json!({"ok": true})),
-            None,
-            None,
-        );
-        bus.set(STATE_SCOPE, &key, terminal).await.unwrap();
-
-        let reply = handle_intercept(&bus, STATE_SCOPE, &call, 1_000, 60_000, false).await;
-        assert_eq!(reply["block"], json!(true));
-        assert_eq!(reply["status"], json!("executed"));
-        // Replay reply: status carries the prior outcome, `replay` discriminator
-        // says we're echoing rather than denying afresh, and no `denial` is
-        // synthesized (the historical record is the source of truth).
-        assert_eq!(reply["replay"], json!("already_resolved"));
-        assert!(reply.get("denial").is_none());
-        assert!(reply.get("reason").is_none());
-
-        // Crucial: the stored row is still `executed`, not overwritten.
-        let stored = bus.get(STATE_SCOPE, &key).await.unwrap();
-        assert_eq!(stored["status"], json!("executed"));
-        assert_eq!(stored["result"], json!({"ok": true}));
-    }
-
-    #[tokio::test]
-    async fn handle_intercept_replay_of_pending_record_preserves_expires_at() {
-        // mutant L331: same branch, pending side. New pending must not bump
-        // the expires_at on the existing row.
-        let bus = InMemoryStateBus::new();
-        let call = sample_call();
-        let key = pending_key(&call.session_id, &call.function_call_id);
-        let pending = build_pending_record(
-            &call.function_call_id,
-            &call.function_id,
-            &call.args,
-            0,
-            60_000,
-        );
-        bus.set(STATE_SCOPE, &key, pending.clone()).await.unwrap();
-
-        let _ = handle_intercept(&bus, STATE_SCOPE, &call, 999_000, 60_000, false).await;
-        let stored = bus.get(STATE_SCOPE, &key).await.unwrap();
-        assert_eq!(
-            stored["expires_at"], pending["expires_at"],
-            "replay must not bump expires_at on the live row"
-        );
-    }
-
-    #[tokio::test]
-    async fn handle_lookup_record_rejects_when_only_one_id_is_empty() {
-        // mutant L395: `||` → `&&` would let one-empty slip through.
-        let bus = InMemoryStateBus::new();
-        let v1 = handle_lookup_record(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "", "function_call_id": "c"}),
-        )
-        .await;
-        assert!(v1.is_null());
-        let v2 = handle_lookup_record(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s", "function_call_id": ""}),
-        )
-        .await;
-        assert!(v2.is_null());
-    }
-
-    #[tokio::test]
-    async fn handle_resolve_rejects_when_only_one_id_is_empty() {
-        // mutant L489: same `||` pattern in handle_resolve guard.
-        let bus = InMemoryStateBus::new();
-        let exec = FakeExecutor::default();
-        let r1 = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &empty_policy_rules(),
-            json!({"session_id": "", "function_call_id": "c", "decision": "allow"}),
-            0,
-        )
-        .await;
-        assert_eq!(r1["error"], json!("missing_id"));
-        let r2 = handle_resolve(
-            &bus,
-            &exec,
-            STATE_SCOPE,
-            &empty_policy_rules(),
-            json!({"session_id": "s", "function_call_id": "", "decision": "allow"}),
-            0,
-        )
-        .await;
-        assert_eq!(r2["error"], json!("missing_id"));
-    }
-
-    #[tokio::test]
-    async fn handle_ack_delivered_returns_zero_when_only_one_field_is_empty() {
-        // mutant L677: two `||` operators in the empty-field guard.
-        let bus = InMemoryStateBus::new();
-        // empty turn_id
-        let r1 = handle_ack_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s", "turn_id": "", "call_ids": ["c"]}),
-        )
-        .await;
-        assert_eq!(r1["stamped"], json!(0));
-        // empty call_ids
-        let r2 = handle_ack_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s", "turn_id": "t", "call_ids": []}),
-        )
-        .await;
-        assert_eq!(r2["stamped"], json!(0));
-        // empty session_id
-        let r3 = handle_ack_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "", "turn_id": "t", "call_ids": ["c"]}),
-        )
-        .await;
-        assert_eq!(r3["stamped"], json!(0));
-    }
-
-    #[test]
-    fn collect_timed_out_for_sweep_rejects_record_missing_only_call_id() {
-        // mutant L423: `||` → `&&` would let one-empty records sweep.
-        let mut rec = build_pending_record("c1", "shell::fs::write", &json!({}), 0, 60_000);
-        rec.as_object_mut()
-            .unwrap()
-            .insert("session_id".into(), json!("s1"));
-        rec.as_object_mut()
-            .unwrap()
-            .insert("function_call_id".into(), json!(""));
-        let out = collect_timed_out_for_sweep(&[rec], 70_000);
-        assert!(out.is_empty(), "empty function_call_id must skip sweep");
-    }
-
-    #[tokio::test]
-    async fn handle_intercept_replay_of_approved_record_preserves_state() {
-        // mutant L331:42 — replace `==` with `!=` on the "approved" side.
-        // The L331:19 mutation is killed by the *_pending_* test above;
-        // this one requires an approved record specifically.
-        let bus = InMemoryStateBus::new();
-        let call = sample_call();
-        let key = pending_key(&call.session_id, &call.function_call_id);
-        let approved = transition_record(
-            &build_pending_record(
-                &call.function_call_id,
-                &call.function_id,
-                &call.args,
-                0,
-                60_000,
-            ),
-            "approved",
-            None,
-            None,
-            None,
-        );
-        bus.set(STATE_SCOPE, &key, approved.clone()).await.unwrap();
-
-        let _ = handle_intercept(&bus, STATE_SCOPE, &call, 999_000, 60_000, false).await;
-        let stored = bus.get(STATE_SCOPE, &key).await.unwrap();
-        assert_eq!(
-            stored["status"],
-            json!("approved"),
-            "replay of approved row must keep status; mutant would overwrite with pending"
-        );
-    }
-
-    #[tokio::test]
-    async fn handle_lookup_record_short_circuits_before_bus_get_on_one_empty_id() {
-        // mutant L395 — `||` → `&&` would let one-empty slip into bus.get.
-        // Seed a record at the address the mutant would compute (pending_key("", "c") = "/c"),
-        // so the mutant returns the seeded row while original code stays at Null.
-        let bus = InMemoryStateBus::new();
-        bus.set(STATE_SCOPE, "/c", json!({"sentinel": "should_not_leak"}))
-            .await
-            .unwrap();
-        let v = handle_lookup_record(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "", "function_call_id": "c"}),
-        )
-        .await;
-        assert!(
-            v.is_null(),
-            "must short-circuit; the seeded sentinel must not leak through"
-        );
-    }
-
-    #[tokio::test]
-    async fn handle_ack_delivered_short_circuits_before_stamping_on_one_empty_field() {
-        // mutant L677 — two `||` operators. If either flips to `&&`, the
-        // function falls through and stamps a record even when a required
-        // field is empty. Seed a record so the stamping path can be
-        // observed.
-        let bus = InMemoryStateBus::new();
-        let terminal = transition_record(
-            &build_pending_record("c", "shell::fs::write", &json!({}), 0, 60_000),
-            "executed",
-            Some(json!({"ok": true})),
-            None,
-            None,
-        );
-        bus.set(STATE_SCOPE, &pending_key("s", "c"), terminal)
-            .await
-            .unwrap();
-
-        // empty turn_id — must NOT stamp the seeded record.
-        let r = handle_ack_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s", "turn_id": "", "call_ids": ["c"]}),
-        )
-        .await;
-        assert_eq!(r["stamped"], json!(0));
-        let stored = bus.get(STATE_SCOPE, &pending_key("s", "c")).await.unwrap();
-        assert!(
-            stored.get("delivered_in_turn_id").is_none(),
-            "must not stamp when turn_id is empty; mutant would stamp"
-        );
-
-        // empty call_ids — same property.
-        let r = handle_ack_delivered(
-            &bus,
-            STATE_SCOPE,
-            json!({"session_id": "s", "turn_id": "t", "call_ids": []}),
-        )
-        .await;
-        assert_eq!(r["stamped"], json!(0));
-        let stored = bus.get(STATE_SCOPE, &pending_key("s", "c")).await.unwrap();
-        assert!(
-            stored.get("delivered_in_turn_id").is_none(),
-            "must not stamp when call_ids is empty"
-        );
-    }
-
-    #[test]
-    fn maybe_flip_timed_out_flips_at_exact_expires_at() {
-        // mutant L439: `<` → `<=` would not flip at the exact boundary.
-        let rec = build_pending_record("c1", "f", &json!({}), 0, 60_000);
-        // expires_at = 0 + 60_000 = 60_000. At now=60_000 the gate
-        // considers the record expired (strictly past or AT expiry).
-        assert!(
-            maybe_flip_timed_out(&rec, 60_000).is_some(),
-            "must flip at exactly expires_at"
-        );
-        assert!(
-            maybe_flip_timed_out(&rec, 59_999).is_none(),
-            "must not flip one ms before expires_at"
-        );
-    }
-
-    // ── proptest: state-machine invariants ────────────────────────────────
-    //
-    // Random sequences of intercept/resolve/sweep/ack/lazy-flip operations
-    // on a single (session, call) record. After every step we assert four
-    // invariants that the lifecycle is supposed to guarantee:
-    //
-    //   I1. status ∈ {pending, approved, executed, failed, denied, timed_out}.
-    //       Any other string is a corrupt record.
-    //   I2. Once a terminal status is observed, the record never returns to
-    //       `pending`. Terminal = executed | failed | denied | timed_out.
-    //   I3. Every `pending` record carries an `expires_at: u64`. Without it
-    //       the sweeper and lazy-flip paths can't classify the record.
-    //   I4. `delivered_in_turn_id` is monotonic: once a non-null value is
-    //       written it is never unset, never replaced with a different turn.
-    //
-    // If any future change can produce a sequence that violates one of
-    // these, proptest will shrink to the minimal failing sequence and
-    // surface it as a counterexample.
-
-    use proptest::prelude::*;
-
-    #[derive(Debug, Clone)]
-    enum Op {
-        InterceptRequired,
-        InterceptNotRequired,
-        ResolveAllow,
-        ResolveDeny,
-        AdvanceClockAndLazyFlip, // bumps clock past expires_at, hits list_undelivered
-        SweepSession,
-        AckDelivered,
-    }
-
-    fn arb_op() -> impl Strategy<Value = Op> {
-        prop_oneof![
-            Just(Op::InterceptRequired),
-            Just(Op::InterceptNotRequired),
-            Just(Op::ResolveAllow),
-            Just(Op::ResolveDeny),
-            Just(Op::AdvanceClockAndLazyFlip),
-            Just(Op::SweepSession),
-            Just(Op::AckDelivered),
-        ]
-    }
-
-    fn make_call(approval_required_self: bool) -> IncomingCall {
-        IncomingCall {
-            session_id: "s".into(),
-            function_call_id: "c".into(),
-            function_id: "test::write".into(),
-            args: json!({}),
-            approval_required: if approval_required_self {
-                vec!["test::write".into()]
-            } else {
-                vec!["other::fn".into()]
-            },
-            event_id: "e".into(),
-            reply_stream: "r".into(),
-        }
-    }
-
-    proptest! {
-        #![proptest_config(ProptestConfig {
-            cases: 256,
-            .. ProptestConfig::default()
-        })]
-
-        #[test]
-        fn state_machine_invariants(ops in prop::collection::vec(arb_op(), 1..30)) {
-            let rt = tokio::runtime::Builder::new_current_thread()
-                .enable_all()
-                .build()
-                .expect("tokio runtime");
-
-            rt.block_on(async {
-                let bus = InMemoryStateBus::new();
-                let exec = FakeExecutor::default();
-                let session_id = "s";
-                let call_id = "c";
-                let timeout_ms: u64 = 60_000;
-                let mut now_ms: u64 = 1_000;
-
-                let mut ever_terminal = false;
-                let mut last_delivered: Option<String> = None;
-
-                for op in &ops {
-                    match op {
-                        Op::InterceptRequired => {
-                            let call = make_call(true);
-                            let _ = handle_intercept(&bus, STATE_SCOPE, &call, now_ms, timeout_ms, false).await;
-                        }
-                        Op::InterceptNotRequired => {
-                            let call = make_call(false);
-                            let _ = handle_intercept(&bus, STATE_SCOPE, &call, now_ms, timeout_ms, false).await;
-                        }
-                        Op::ResolveAllow => {
-                            let _ = handle_resolve(
-                                &bus,
-                                &exec,
-                                STATE_SCOPE,
-                                &empty_policy_rules(),
-                                json!({
-                                    "session_id": session_id,
-                                    "function_call_id": call_id,
-                                    "decision": "allow",
-                                }),
-                                now_ms,
-                            )
-                            .await;
-                        }
-                        Op::ResolveDeny => {
-                            let _ = handle_resolve(
-                                &bus,
-                                &exec,
-                                STATE_SCOPE,
-                                &empty_policy_rules(),
-                                json!({
-                                    "session_id": session_id,
-                                    "function_call_id": call_id,
-                                    "decision": "deny",
-                                }),
-                                now_ms,
-                            )
-                            .await;
-                        }
-                        Op::AdvanceClockAndLazyFlip => {
-                            now_ms = now_ms.saturating_add(timeout_ms + 1);
-                            let _ = handle_list_undelivered(
-                                &bus, STATE_SCOPE,
-                                json!({ "session_id": session_id }),
-                                now_ms,
-                            ).await;
-                        }
-                        Op::SweepSession => {
-                            let _ = handle_sweep_session(
-                                &bus, STATE_SCOPE,
-                                json!({ "session_id": session_id }),
-                            ).await;
-                        }
-                        Op::AckDelivered => {
-                            let _ = handle_ack_delivered(
-                                &bus, STATE_SCOPE,
-                                json!({
-                                    "session_id": session_id,
-                                    "turn_id": format!("turn-{now_ms}"),
-                                    "call_ids": [call_id],
-                                }),
-                            ).await;
-                        }
-                    }
-
-                    // Assert invariants on whatever the record currently is.
-                    let key = pending_key(session_id, call_id);
-                    let Some(rec) = bus.get(STATE_SCOPE, &key).await else {
-                        // No record yet (e.g. only InterceptNotRequired so far). Skip.
-                        continue;
-                    };
-
-                    // I1: legal status
-                    let status = rec.get("status").and_then(Value::as_str).unwrap_or("");
-                    assert!(
-                        matches!(
-                            status,
-                            "pending" | "approved" | "executed" | "failed" | "denied" | "timed_out"
-                        ),
-                        "I1 violated: illegal status {status:?} after ops {ops:?}; record={rec:?}"
-                    );
-
-                    // I2: no reverting terminal → pending
-                    if matches!(status, "executed" | "failed" | "denied" | "timed_out") {
-                        ever_terminal = true;
-                    }
-                    if ever_terminal {
-                        assert!(
-                            status != "pending",
-                            "I2 violated: reverted to pending after terminal; ops={ops:?}; record={rec:?}"
-                        );
-                    }
-
-                    // I3: pending records always have expires_at: u64
-                    if status == "pending" {
-                        let exp = rec.get("expires_at").and_then(Value::as_u64);
-                        assert!(
-                            exp.is_some(),
-                            "I3 violated: pending record missing expires_at; ops={ops:?}; record={rec:?}"
-                        );
-                    }
-
-                    // I4: delivered_in_turn_id is monotonic — once set non-null, never unset / never replaced
-                    let cur_delivered = rec
-                        .get("delivered_in_turn_id")
-                        .and_then(Value::as_str)
-                        .map(str::to_string);
-                    if let Some(prev) = &last_delivered {
-                        match &cur_delivered {
-                            Some(cur) => {
-                                assert_eq!(
-                                    cur, prev,
-                                    "I4 violated: delivered_in_turn_id replaced {prev:?} → {cur:?}; ops={ops:?}"
-                                );
-                            }
-                            None => {
-                                panic!(
-                                    "I4 violated: delivered_in_turn_id unset after being {prev:?}; ops={ops:?}; record={rec:?}"
-                                );
-                            }
-                        }
-                    }
-                    if cur_delivered.is_some() {
-                        last_delivered = cur_delivered;
-                    }
-                }
-            });
-        }
-    }
-}
